Squashed commit of the following:
commit 0a451c9e9e881a0f534d827a9f35ff77f909d295
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Wed Aug 9 23:12:12 2023 +0900

    for mac

    mac

commit 82198929059b57e272335b6b2a123bd6a42ff1d5
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Wed Aug 9 20:51:05 2023 +0900

    멀까

    ?

commit 2ee12f0b04d9fc2f410f5579d63c6aec79d9d58b
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Wed Aug 9 19:18:47 2023 +0900

    ??

    ??

commit cc58561810741db7cdcb621a348488d25fdce62c
Merge: 573531b fd3943d
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Wed Aug 9 16:30:08 2023 +0900

    Merge branch 'develop' into feature_project_search_fix

commit 573531b6a746c82a346799a6d53ac1f98397cbef
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Wed Aug 9 15:46:22 2023 +0900

    권한확인 기능 추가

    type:change
    body : 권한확인이 필요한 기능에 권한확인 기능 적용

commit 474cae3d24d83cfb7c142473b7ad7f0836089526
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Tue Aug 8 16:28:46 2023 +0900

    검색기능 버그 수정

    type : fix
    body : 검색 기능 버그 수정

commit d37039083c8842ec0762351e2f9a3bf3da171373
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Tue Aug 8 15:35:13 2023 +0900

    Feature:getRole

    유저역할을 확인하는 기능 추가

commit d49f5dc1790b8f1f25984a2633439daa772baeb5
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Mon Jul 31 09:49:40 2023 +0900

    0731 최신화

    DTO일부 수정

commit 16e993100b8cd42f7e6272761e3e8cf6c04c449f
Author: cmilk18 <43445697+cmilk18@users.noreply.github.com>
Date:   Wed Jul 26 17:17:53 2023 +0900

    0726 hotfix 검색

    검색

+++ src/main/java/com/swave/urnr/project/service/ProjectServiceImpl.java
@@ -2,6 +2,7 @@ package com.swave.urnr.project.service;
 
 
 import com.swave.urnr.project.domain.Project;
+import com.swave.urnr.project.exception.NotAuthorizedException;
 import com.swave.urnr.project.repository.ProjectRepository;
 import com.swave.urnr.project.requestdto.ProjectCreateRequestDTO;
 import com.swave.urnr.project.requestdto.ProjectKeywordRequestContentDTO;
@@ -13,6 +14,7 @@ import com.swave.urnr.user.domain.UserInProject;
 import com.swave.urnr.user.repository.UserInProjectRepository;
 import com.swave.urnr.user.repository.UserRepository;
 import com.swave.urnr.user.responsedto.UserMemberInfoResponseDTO;
+import com.swave.urnr.user.service.UserInProjectService;
 import com.swave.urnr.util.type.UserRole;
 import lombok.RequiredArgsConstructor;
 import com.swave.urnr.util.http.HttpResponse;
@@ -21,16 +23,19 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 import javax.sql.DataSource;
 
-
+import java.io.IOException;
 import java.io.UnsupportedEncodingException;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
+
+import java.util.*;
+
 
 ;import static com.swave.urnr.project.domain.Project.makeProjectSearchListResponseDTOList;
 import static com.swave.urnr.util.type.UserRole.Manager;
+import static com.swave.urnr.util.type.UserRole.None;
 
 //0710 확인 CR
 @Service
@@ -44,6 +49,7 @@ public class ProjectServiceImpl implements ProjectService {
     private final UserRepository userRepository;
 
     private final ReleaseNoteRepository releaseNoteRepository;
+
 
     @Override
     @Transactional
@@ -135,45 +141,12 @@ public class ProjectServiceImpl implements ProjectService {
                 .build();
     }
 
-    //requestGetAttribute
-    //서블렛
-    /*@Override
-    public String updateUsers(ProjectRequestDto projectRequestDto) {
-        //프로젝트를 불러와서
-        //리스트 받아서
-        //유저인리스트 와
-        //유저에 매핑
-
-        //모든객체 싹다 호출
-        //Project project = projectRepository.findById(projectRequestDto.getId()).orElse(null);
-        //List<UserInProject> teamMembers = new ArrayList<>(project.getUserInProjectList());
-
-
-        for(Long userId : projectRequestDto.getUsers() )
-        {
-            UserInProject userInProject = new UserInProject();
-            userInProject.setRole(UserRole.Developer);
-            userInProject.setProject(project);
-
-            User user = userRepository.findById(userId).get();
-            //User user1 = userRepository.findById(userInProject1.getId()).get();
-            userInProject.setUser(user);
-            userInProjectRepository.save(userInProject);
-            List<UserInProject> userInProjectList = user.getUserInProjectList();
-            userInProjectList.add(userInProject);
-            user.setUserInProjectList(userInProjectList);
-            userRepository.flush();
-
-            project.getUserInProjectList().add(userInProject);
-        }
-        projectRepository.flush();
-        return null;
-    }*/
-
     //최신 릴리즈노트 이름 가져오기
     @Override
     @Transactional(readOnly = true)
     public List<ProjectListResponseDTO> loadProjectList(HttpServletRequest request) {
+
+
         List<ProjectListResponseDTO> projectList = new ArrayList<>();
         List<UserInProject> userInProjectList = userInProjectRepository.findByUser_Id((Long)request.getAttribute("id"));
 
@@ -199,6 +172,7 @@ public class ProjectServiceImpl implements ProjectService {
     @Override
     @Transactional(readOnly = true)
     public ProjectContentResponseDTO loadProject(Long projectId) {
+
         Project project = projectRepository.findById(projectId).get();
         ProjectContentResponseDTO getproject = ProjectContentResponseDTO.builder()
                 .id(project.getId())
@@ -222,10 +196,28 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     @Override
-    @Transactional(readOnly = true)
-    public ProjectManagementContentResponseDTO loadManagementProject(HttpServletRequest request,Long projectId) {
-        //유저가 해당 프로젝트 멤버인지 확인 UserInList확인
-        //유저가 관리자인지 확인 UserInList확인
+    public UserRole getRole(HttpServletRequest request, Long projectId) {
+        UserInProject userInProject = userInProjectRepository.findByUser_IdAndProject_Id((Long)request.getAttribute("id"),projectId);
+
+        UserRole role;
+        if(userInProject==null){
+            role = None;
+        }else{
+            role = userInProject.getRole();
+        }
+
+        return role;
+    }
+
+    @Override
+    public ProjectManagementContentResponseDTO loadManagementProject(HttpServletRequest request,Long projectId) throws NotAuthorizedException {
+        //todo:권한체크
+        UserRole role = getRole(request, projectId);
+        if (role != UserRole.Manager) {
+            throw new NotAuthorizedException("해당 프로젝트의 매니저만 접근할 수 있습니다.");
+        }
+
+
         Project project = projectRepository.findById(projectId).get();
         User user = userRepository.findById((Long)request.getAttribute("id")).orElse(null);
         List<UserMemberInfoResponseDTO> getMembers = userInProjectRepository.getMembers(projectId);
@@ -251,10 +243,17 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Override
     @Transactional(readOnly = true)
-    public ProjectManagementContentResponseDTO loadManagementProjectJPA(HttpServletRequest request, Long projectId){
+
+    public ProjectManagementContentResponseDTO loadManagementProjectJPA(HttpServletRequest request, Long projectId) throws NotAuthorizedException {
+
         Project project = projectRepository.findById(projectId).get();
         User user = userRepository.findById((Long)request.getAttribute("id")).orElse(null);
         List<UserMemberInfoResponseDTO> getMembers  = new ArrayList<>();
+        //todo:권한체크
+        UserRole role = getRole(request, projectId);
+        if (role != UserRole.Manager) {
+            throw new NotAuthorizedException("해당 프로젝트의 매니저만 접근할 수 있습니다.");
+        }
 
         for(UserInProject userInProject:project.getUserInProjectList()){
             UserMemberInfoResponseDTO userMemberInfoResponseDTO = new UserMemberInfoResponseDTO();
@@ -281,7 +280,13 @@ public class ProjectServiceImpl implements ProjectService {
     //관리자 변경하기 의외로 쉽게 될수도?
     @Override
     @Transactional
-    public ProjectUpdateRequestDTO updateProject(Long projectId, ProjectUpdateRequestDTO projectUpdateRequestDto) {
+    public ProjectUpdateRequestDTO updateProject(HttpServletRequest request, Long projectId, ProjectUpdateRequestDTO projectUpdateRequestDto) throws NotAuthorizedException {
+
+        //todo:권한체크
+        UserRole role = getRole(request, projectId);
+        if (role != UserRole.Manager) {
+            throw new NotAuthorizedException("해당 프로젝트의 매니저만 접근할 수 있습니다.");
+        }
 
         System.out.println(projectId);
         Project project = projectRepository.findById(projectId).orElseThrow(null);
@@ -326,7 +331,15 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Override
     @Transactional
-    public HttpResponse deleteProject(Long projectId) {
+
+    public HttpResponse deleteProject(HttpServletRequest request, Long projectId) throws NotAuthorizedException {
+
+        //todo:권한체크
+        UserRole role = getRole(request, projectId);
+        if (role != UserRole.Manager) {
+            throw new NotAuthorizedException("해당 프로젝트의 매니저만 접근할 수 있습니다.");
+        }
+
         List<UserInProject> userInProjectList = userInProjectRepository.findByProject_Id(projectId);
         for(UserInProject userInProject:userInProjectList) {
             userInProjectRepository.delete(userInProject);
@@ -342,6 +355,9 @@ public class ProjectServiceImpl implements ProjectService {
     @Override
     @Transactional(readOnly = true)
     public ProjectSearchResultListResponseDTO searchProject(ProjectKeywordRequestContentDTO projectKeywordRequestContentDTO) throws UnsupportedEncodingException {
+
+        //todo:권한체크
+
         String keyword = projectKeywordRequestContentDTO.getKeyword();
         ProjectSearchResultListResponseDTO projectSearchResultListResponseDTO = new ProjectSearchResultListResponseDTO();
